Make zone search a bit more useful
Allow "-list" without a filter argument: this lists all the zones we
know of, and makes fuzzy searching with programs like fzf possible.

When -list *is* followed by another argument, tz will try to match on
the zone name, or abbreviated zone name in a case insensitive manner.

So far, -list results are sorted alphabetically. I think it could make
more sense to use the offset, but YMMV.